+++ resources/views/admin/categories/create.blade.php
@@ -1,0 +1,16 @@
+@extends('admin.master')
+
+@section('main_content')
+    <div class="row">
+        <div class="col-md-8 col-md-offset-2">
+            <h1>Create New Category</h1>
+            <hr>
+            <form class="form-horizontal">
+                <label class="control-label">Test:</label>
+                <select name="name" class="form-control">
+                    <option value="">Salam</option>
+                </select>
+            </form>
+        </div>
+    </div>
+@stop
