+++ resources/views/admin/categories/manageChild.blade.php
@@ -3,7 +3,7 @@
    <li>
       {{ $child->name }}
       @if(count($child->childs))
-      @include('admin.manageChild',['childs' => $child->childs])
+      @include('admin.categories.manageChild',['childs' => $child->childs])
       @endif
    </li>
    @endforeach
